Begin large-scale migration of command procedure code
